refactor: separate agent query hooks and check access to enable fetching
